pyproject.toml
[project]
name = "ha-dwelo"
version = "0.1.0"
description = "Custom Home Assistant integration for Dwelo smart apartments"
requires-python = ">=3.12"
dependencies = [
    "aiohttp>=3.9.0",
]

[dependency-groups]
dev = [
    "playwright>=1.58.0",
    "pytest>=9.0.2",
    "pytest-asyncio>=1.3.0",
    "python-dotenv>=1.2.2",
]

[tool.pytest.ini_options]
asyncio_mode = "auto"
